// projects/codx-share/src/lib/components/setting-calendar/models/BS_DaysOff.model.ts
export class BS_DaysOff {
  recID: string;
  dayoffCode: string;
  month: string;
  day: string;
  months: any;
  days: any;
  calendar: string;
  isLastYear: boolean;
  specialDay: boolean;
  symbol: string;
  color: string;
  workdayID: string;
  note: string;
  createdOn: string;
  createdBy: string;
  owner: string;
  bUID: string;
  employeeID: string;
  positionID: string;
  orgUnitID: string;
  divisionID: string;
}